dto: complete for all controllers

+++ financial-dashboard-node/src/controllers/budget.controller.ts
@@ -1,12 +1,17 @@
 import { Request, Response, NextFunction } from "express";
 import { BudgetService } from "../services/budget.service";
 import { BudgetCategory } from "../enums/budget.enum";
+import { RequestBudgetDto } from "../dtos/budget/request-budget.dto";
+import { AppError } from "../types/app-error";
+import { AddBudgetDto } from "../dtos/budget/add.dto";
+import { UpdateBudgetDto } from "../dtos/budget/update.dto";
+import { UpdateExpenseDto } from "../dtos/expense/update-expense.dto";
 
 const budgetService = new BudgetService();
 export const getAllBudgets = async(req:Request,res:Response,next:NextFunction)=>{
     try{
         let budget = req.body;
-        budget.amount = parseFloat(budget.amount);
+        // budget.amount = parseFloat(budget.amount);
         let user = req.body.user;
 
         let results = await budgetService.getAllBudgets(user);
@@ -25,22 +30,36 @@ export const getAllBudgets = async(req:Request,res:Response,next:NextFunction)=>
 export const addBudget = async(req:Request,res:Response,next:NextFunction)=>{
     try{   
         let user = req.body.user;
-        let budget = req.body;
+        let {startDate,endDate,...body} = req.body;
+        console.log(body);
         
-        let results = await budgetService.createBudget(budget,user);
+        
+        let error = await AddBudgetDto.validate(body);
 
-        if(results){
-            res.status(200).json({
-                status:'successfull',
-                data:'Budget is added'
-            })
+        if(!error.isValid){
+            throw new AppError('Add Budget data invalid',500);
         }
         else{
-            res.status(400).json({
-                status:'failed',
-                data:'Budget is not added'
-            })
+            let budget = new AddBudgetDto(body);
+            let results = await budgetService.createBudget(budget,user);
+
+            if(results){
+                res.status(200).json({
+                    status:'successfull',
+                    data:'Budget is added'
+                })
+            }
+            else{
+                res.status(400).json({
+                    status:'failed',
+                    data:'Budget is not added'
+                })
+            }
+            // console.log('valid data');
+            
         }
+        
+        
     }
     catch(err){
         next(err);
@@ -158,13 +177,23 @@ export const getTotalSpendingOfCategory = async(req:Request,res:Response,next:Ne
 export const deleteBudget = async(req:Request,res:Response,next:NextFunction)=>{
     try{
         let user = req.body.user;
-        let id = req.params.id;
-        let result = await budgetService.deleteBudget(user,id);
+        let id = parseInt(req.params.id);
+        // console.log(id);
 
-        res.status(200).json({
-            status:'successfull',
-            data:'budget is deleted'
-        })
+        if(!isNaN(id)){
+            let result = await budgetService.deleteBudget(user,id);
+
+            res.status(200).json({
+                status:'successfull',
+                data:'budget is deleted'
+            })
+            console.log("valid");
+            
+        }
+        else{
+            throw new AppError('ID should be number',500);
+        }
+        
     }
     catch(err){
         next(err);
@@ -174,23 +203,31 @@ export const deleteBudget = async(req:Request,res:Response,next:NextFunction)=>{
 
 export const getBudgetById = async(req:Request,res:Response,next:NextFunction)=>{
     try{
-        let id = req.params.id;
         let user = req.body.user;
+        let id = parseInt(req.params.id);
+        // console.log(id);
 
-        let result = await budgetService.getBudgetById(id);
+        if(!isNaN(id)){
+            let result = await budgetService.getBudgetById(id);
 
-        if(result){
-            res.status(200).json({
-                status:'successfull',
-                data:result
-            })
+            if(result){
+                res.status(200).json({
+                    status:'successfull',
+                    data:result
+                })
+            }
+            else{
+                res.status(400).json({
+                    status:'failed',
+                    data:'no budget found'
+                })
+            }
         }
         else{
-            res.status(400).json({
-                status:'failed',
-                data:'no budget found'
-            })
+            throw new AppError('ID should be number',500);
         }
+
+        
     }
     catch(err){
         next(err);
@@ -200,16 +237,39 @@ export const getBudgetById = async(req:Request,res:Response,next:NextFunction)=>
 
 export const updateBudgetById = async(req:Request,res:Response,next:NextFunction)=>{
     try{
-        let budget = req.body;
         let user = req.body.user;
-        let id = req.params.id;
+        let {startDate,endDate,...body} = req.body;
+        // body.amount = Number(body.amount);
+        let id = parseInt(req.params.id);
 
-        let result = await budgetService.updateBudgetById(budget,id);
+        
+        
 
-        res.status(200).json({
-            status:'successfull',
-            data:'budget updated'
-        })
+        if(!isNaN(id)){
+            body.id = id;
+            console.log('Body: ');
+            
+            console.log(body);
+            let error = await UpdateBudgetDto.validate({id:body.id,amount:body.amount});
+
+            if(!error.isValid){
+                throw new AppError('Budget data invalid',500);
+            }
+            else{
+                let budget = new UpdateBudgetDto(body);
+                let result = await budgetService.updateBudgetById(budget,budget.id);
+
+                res.status(200).json({
+                    status:'successfull',
+                    data:'budget updated'
+                })
+                // console.log("valid");
+                
+            }
+        }
+        else{
+            throw new AppError('ID should be number',500);
+        }
     }
     catch(err){
         next(err);
